Fix #2092883: Don't allow manual sync if an item is disabled.
This is a temporary fix to prevent the user from inadvertently syncing an item that has sync disabled.
The use case is if a user intentionally doesn't ever want Contacts synced but has turned off background data or
automatic sync. Instead, we post a dialog informing the user to turn background and auto sync back on to change their
preference.